Finish merge of users and fix database for the new version
