+++ main/views.py
@@ -668,116 +668,3 @@ def profile(request, username):
         plant_slots = None
 
     return render(request, "profile.html", {"owner": username, "plant_slots": plant_slots})
-
-@login_required()
-def settings(request):
-    user_form = ProfileUpdateForm(instance=request.user)
-    password_form = CustomPasswordChangeForm(user=request.user)
-
-    if request.method == 'POST':
-        if 'update_profile' in request.POST:
-            user_form = ProfileUpdateForm(request.POST, instance=request.user)
-            if user_form.is_valid():
-                user_form.save()
-                messages.success(request, 'Your profile has been updated successfully!')
-                return redirect('settings')
-        elif 'change_password' in request.POST:
-            password_form = CustomPasswordChangeForm(user=request.user, data=request.POST)
-            if password_form.is_valid():
-                password_form.save()
-                messages.success(request, 'Your password has been changed successfully!')
-                return redirect('/auth/login/')
-    else:
-        user_form = ProfileUpdateForm(instance=request.user)
-        password_form = CustomPasswordChangeForm(user=request.user)
-
-    context = {
-        'user_form': user_form,
-        'password_form': password_form
-    }
-    return render(request, "settings.html", context)
-
-@login_required
-def delete_account(request):
-    if request.method == 'POST':
-        user = request.user
-        logout(request)
-        user.delete()
-        messages.success(request, 'Your account has been deleted successfully.')
-        return redirect('home')
-    return render(request, 'settings.html')
-
-
-@login_required(login_url="/auth/login")
-def friends(request):
-    user_friends = request.user.get_friends()
-    friends_total = user_friends.count()
-
-    if friends_total > 4:
-        paginator = Paginator(user_friends, 4)
-        page_number = request.GET.get('page')
-        friends_page = paginator.get_page(page_number)
-    else:
-        friends_page = user_friends
-
-    return render(request, "friends.html", {"friends_page": friends_page, "friends_total": friends_total})
-@login_required
-def remove_friend(request):
-    if request.method == "POST":
-        data = json.loads(request.body)
-        friend_id = data.get("user_id")
-        if not friend_id:
-            return HttpResponse("Invalid request", status=400)
-        friend = get_object_or_404(CustomUser, id=friend_id)
-        request.user.unfriend(friend)
-        return HttpResponse(status=200)
-    return HttpResponse("Invalid request method", status=405)
-
-
-@login_required
-def send_friend_request(request):
-    if request.method == "POST":
-        try:
-            data = json.loads(request.body)
-            username = data.get("username")
-            if not username:
-                return JsonResponse({"success": False, "message": "Username is required."})
-            try:
-                to_user = CustomUser.objects.get(username=username)
-            except CustomUser.DoesNotExist:
-                return JsonResponse({"success": False, "message": "User does not exist."})
-            request.user.send_friend_request(to_user)
-            return JsonResponse({"success": True, "message": "Friend request sent."})
-        except ValueError as e:
-            return JsonResponse({"success": False, "message": str(e)})
-
-
-@login_required
-def accept_friend_request(request):
-    if request.method == "POST":
-        data = json.loads(request.body)
-        friend_id = data.get("user_id")
-        if not friend_id:
-            return HttpResponse("Invalid request.", status=400)
-        friend = get_object_or_404(CustomUser, id=friend_id)
-        request.user.accept_friend_request(friend)
-        return HttpResponse(status=200)
-    return HttpResponse("Invalid request method", status=405)
-
-
-@login_required
-def reject_friend_request(request):
-    if request.method == "POST":
-        data = json.loads(request.body)
-        friend_id = data.get("user_id")
-        if not friend_id:
-            return HttpResponse("Invalid request.", status=400)
-        friend = get_object_or_404(CustomUser, id=friend_id)
-        request.user.reject_friend_request(friend)
-        return HttpResponse(status=200)
-    return HttpResponse("Invalid request method", status=405)
-
-@login_required
-def get_friend_requests(request):
-    friend_requests = request.user.get_incoming_friend_requests().values("id", "username")
-    return JsonResponse({"friend_requests": list(friend_requests)})
