+++ users/apps.py
@@ -4,4 +4,7 @@ from django.apps import AppConfig
 class UsersConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'users'
-    
+
+    def ready(self):
+        import users.signals
+
